requirements.txt
certifi==2018.10.15
chardet==3.0.4
cloudant==2.10.0
Django==1.11.16
django-environ==0.4.5
idna==2.7
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
